Create problem_5.c
You have given positive integer N, the size of an array as input. Take an array of size N as input. You have to find the
sum of those element in array which last digit is zero. Implement this by using function.
Note - It is guaranteed that each element of an array is at least 2 digit number.
Sample Input:                             Sample Output:
5                                                   200
110 24 50 35 40